:shirt: split utility functions from kube-interface module #sdk/k8s-operator
